feat: refactor tafsir panel components
