+++ app/(features)/surah/[surahId]/components/tafsir-panel/index.ts
@@ -1,0 +1,6 @@
+export { TafsirPanel } from '../TafsirPanel';
+export { useTafsirPanel } from './useTafsirPanel';
+export { TafsirTabs } from './TafsirTabs';
+export { TafsirList } from './TafsirList';
+export { TafsirItem } from './TafsirItem';
+export * from './tafsirPanel.utils';
