+++ app/(features)/surah/[surahId]/components/tafsir-panel/tafsirPanel.utils.ts
@@ -1,0 +1,15 @@
+export interface Tafsir {
+  id: number;
+  name: string;
+  lang: string;
+  selected: boolean;
+}
+
+export const MAX_SELECTIONS = 3;
+
+export const capitalizeLanguageName = (lang: string): string => {
+  return lang
+    .split(' ')
+    .map((word) => word.charAt(0).toUpperCase() + word.slice(1).toLowerCase())
+    .join(' ');
+};
